backend-inventory/requirements.txt
django>=4.2.7
djangorestframework
django-cors-headers
